Shift to new ingest-utils json parsing method (#76)
Why

The DAP pipeline does not terminate cleanly if an exception is thrown from the Redcap client (usually if non-JSON garbage is returned). We should handle that situation and terminate.

This PR

Shifts to a new json parsing method introduced in ingest-utils 2.1.4 that returns an Either[ParsingFailure, Msg] that we will match against.