[DSPDC-1771] Add explicit inter-solid dependencies (#172)
Why

As part of my spiking for the TDR ticket, I discovered we had no explicit dependencies between solids in our refresh pipeline definition. This means that solids can effectively execute in any order and potentially in parallel. We have not bumped into this in practice because dagster runs the solids in the order in which they're declared and serially, when run locally.

This PR

Introduces explicit solid dependencies between the extract + transform pairs, and a final "fan-in" to the write_outfiles solid. The fan-in consumes a list of results that gets dropped on the floor, but is necessary for the dependency to get created.
Adds a test case that runs the pipeline using mock resources. I have found this sort of test useful to verify the overall pipeline definition, especially when there are complex fan-ins/outs.
Shifts our tests to be pure pytest
Adds api_token, pull_data_dictionaries, and end_time as explicit config schema entries for the extract solids. I am not sure how this ever worked, but I was unable to run the pipeline without these schema items defined.