[DSPDC-1891] Output TSVs to GS paths (#193)
Why

We need to be able to output tsvs to a GS path when running in GCP (as opposed to a local FS path when running locally)

This PR

Refactors the convert-output-to-tsv.py script, placing it in the orchestration tree and making it callable from our orch. code directly, rather than having to use subprocess.run
Threads through the dap survey type from extract -> transform -> write_outfiles so we can run subsets of the pipelines as needed (i.e. just eols and sample for example)